backend/requirements.txt
fastapi>=0.104.1,<0.105.0
uvicorn[standard]>=0.24.0,<0.25.0
python-dotenv>=1.0.0,<2.0.0
snowflake-connector-python>=3.7.0,<3.8.0
passlib[bcrypt]>=1.7.4,<1.8.0
bcrypt>=4.0.0,<5.0.0
python-jose[cryptography]>=3.3.0,<3.4.0
python-multipart>=0.0.6,<0.1.0
reportlab>=4.0.0,<4.1.0
pydantic>=2.5.0,<2.6.0
pydantic-settings>=2.1.0,<2.2.0
httpx>=0.25.0,<0.26.0
sqlparse>=0.4.4,<0.5.0
black>=23.3.0,<24.0.0
isort>=5.12.0,<6.0.0
flake8>=6.0.0,<7.0.0
mypy>=1.3.0,<2.0.0
